refactor: split AdminView into tab components and add modal error handling
